Introduce a struct dcc_ops to contain all functions pointers for a custom DCC type
Modules now must pass a pointer to a struct dcc_ops when calling add_dcc_bind() instead of
a list of function pointers.

+++ dll/arcfour/arcfour.h
@@ -9,6 +9,6 @@ static inline void arcfourInit(arckey *, void *, unsigned short);
 static inline char *arcfourCrypt(arckey *, char *, int);
 static int send_dcc_encrypt (int, int, char *, int);
 static int dcc_schat_input (int, int, char *, int, int);
-static int start_dcc_crypt (int, int, unsigned long, int);
-static int end_dcc_crypt (int, unsigned long, int);
+static int start_dcc_crypt (int, int, unsigned long, unsigned short);
+static int end_dcc_crypt (int, unsigned long, unsigned short);
 void dcc_sdcc (char *, char *);
